+++ fiddles/angular2-cli/fiddle-0023-EntAgGridRowSelection/src/app/ag-tree-grid/interfaces/ag-tree-grid-impl.ts
@@ -1,0 +1,10 @@
+import { AgTreeGridOptions } from './ag-tree-grid-options';
+import { AgTreeGridEvent } from './ag-tree-grid-event';
+
+export interface AgTreeGridImpl {
+
+  options: AgTreeGridOptions;
+
+  agTreeGridOnEvent(event: AgTreeGridEvent): void;
+
+}