+++ fiddles/angular2-cli/fiddle-0025-EntAgGridTreeExpandEvent/src/app/ag-tree-grid/ag-tree-grid.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AgTreeGridComponent } from './ag-tree-grid.component';
+import {AgGridModule} from 'ag-grid-angular/main';
+
+import 'ag-grid-enterprise';
+
+@NgModule({
+  declarations: [AgTreeGridComponent],
+  imports: [
+    CommonModule,
+    AgGridModule.withComponents([])
+  ],
+  exports: [
+    AgTreeGridComponent
+  ]
+})
+export class AgTreeGridModule { }