+++ fiddles/angular2-cli/fiddle-0025-EntAgGridTreeExpandEvent/src/app/ag-tree-grid/interfaces/ag-tree-grid-event.ts
@@ -1,0 +1,9 @@
+import {
+  AgTreeGridEvents
+} from '../enums/index';
+
+export interface AgTreeGridEvent {
+  type: AgTreeGridEvents;
+  data: any;
+  gridState?: any;
+}