+++ fiddles/angular2-cli/fiddle-0025-EntAgGridTreeExpandEvent/src/app/ag-tree-grid/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './ag-tree-grid-event';
+export * from './ag-tree-grid-options';
+export * from './ag-tree-grid-impl';
+