+++ fiddles/angular2-cli/fiddle-0025-EntAgGridTreeExpandEvent/src/app/interfaces/events.ts
@@ -1,0 +1,16 @@
+export interface Events {
+  /**
+   * The type of event.
+   */
+  type?: string | number;
+
+  /**
+   * Data that is emitted.
+   */
+  data?: any;
+
+  /**
+   * grid state -- optional parameter used by ag-grid
+   */
+  gridState?: any;
+}