+++ CHANGELOG.md
@@ -2160,3 +2160,7 @@ Alternately, the _Daily Diary of my Programming Dreams_.
   4. Updated [scripts/fiddle-publish.sh](scripts/fiddle-publish.sh) 
 
 
+* Added [fiddles/angular2-cli/fiddle-0022-EntAgGridTreeData](fiddles/angular2-cli/fiddle-0022-EntAgGridTreeData)
+* Added [fiddles/angular2-cli/fiddle-0023-EntAgGridRowSelection](fiddles/angular2-cli/fiddle-0023-EntAgGridRowSelection)
+* Added [fiddles/angular2-cli/fiddle-0024-EntAgGridRowTransactions](fiddles/angular2-cli/fiddle-0024-EntAgGridRowTransactions)
+* Added [fiddles/angular2-cli/fiddle-0025-EntAgGridTreeExpandEvent](fiddles/angular2-cli/fiddle-0025-EntAgGridTreeExpandEvent)